requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==4.0.4
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
autopep8==2.3.2
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.13.3
black==26.1.0
bleach==6.2.0
cachetools==6.2.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.122.0
fastjsonschema==2.21.1
flake8==7.3.0
flatbuffers==25.12.19
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
gast==0.7.0
google-auth==2.43.0
google-genai==1.52.0
google-pasta==0.2.0
greenlet==3.2.4
groq==0.37.1
grpcio==1.78.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==7.0.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.13.2
kiwisolver==1.4.7
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.44
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.20
mistune==3.1.1
ml_dtypes==0.5.4
MouseInfo==0.1.3
multidict==6.1.0
mypy_extensions==1.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==1.0.4
phonenumbers==9.0.3
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==7.34.1
psutil==6.1.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.12.0
pydantic_core==2.27.2
pyflakes==3.4.0
PyGetWindow==0.0.9
Pygments==2.19.1
pylint==4.0.4
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.2.1
pytokens==0.4.1
pytweening==1.2.0
pytz==2025.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.22.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
tensorflow==2.21.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tk==0.1.0
tomlkit==0.14.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.46.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==2.1.2
xxhash==3.6.0
yapf==0.43.0
yarl==1.18.3
yt-dlp==2025.3.31
zstandard==0.25.0
